requirements.txt
codecov==2.0.15
coverage==4.5.1
matplotlib==2.2.2
numpy==1.14.3
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pytest==3.5.0
pytest-cov==2.5.1
pytest-flake8==0.9.1
